scripts/pyproject.toml
[project]
name = "neewer-scripts"
version = "0.1.0"
description = "Reverse engineering scripts for Neewer light control"
requires-python = ">=3.10"
dependencies = [

    "frida",

    "pefile",

    "pillow",

    "hidapi",

    "pyserial",

    "pyusb>=1.3.1",
]

[tool.setuptools]
py-modules = []

